Updated default asset URL in configuration file
When you have a Laravel application running in a sub-folder, an absolute path will be generated to the root of the URL, and not the sub-folder URL.

For example, I have the following URL (Laravel application inside of the `erm` folder):

http://reports.company.com/erm/

The `<livewire:scripts>` annotiation will generate:

```html
<script src="/vendor/livewire/livewire.js?id=c6f7d3682bc8d847a034" data-turbolinks-eval="false"></script>
```

Which 404's. With the proposed change using the `url()` helper, it is set to the proper sub-folder:

```html
<script src="http://reports.company.com/erm/vendor/livewire/livewire.js?id=c6f7d3682bc8d847a034" data-turbolinks-eval="false"></script>
```

This is simply a change to the default asset URL. This won't break and users current asset URLs set.