Fixes 'Typed property must not be accessed before initialization'
Typed properties are a feature of PHP 7.4, which allows
properties to reference specific types, helping both during
development and runtime.

This new implementation has the need to provide an "initialized" state for
all typed properties, which breaks Livewire's implementation when accesing
properties in early component construction.

The 'ReflectionProperty' class has a new 'isInitialized' method, which
helps us to determine if the value of the property is accessible,
but is only available for PHP >= 7.4.

In order to prevent Livewire's property implementation from breaking
when developers want to use typed properties, we check for the existance
of the 'isInitialized' method in the 'ReflectionProperty' class.
If the method doesn't exist, it must mean that the running PHP version
does not support typed properties, so we just access the property value directly,
expecting a default null value if not initialized, or an actual value if already initialized.

If the method exists, then we return a value depending on the property state.
If the property has not been initialized, we just return a null value (like earlier PHP
versions), otherwise, we access the property value and return it.

This commit closes #707 and closes #600.
There are no breaking changes.