Allow use of custom rule classes in validation
This is a relatively minor change that uses the `rules` method as the sole accessor for the validation rules array.  As the existing method uses a class property, one cannot instantiate a new rule inside it.

```php
class SomeComponent extends Component {
    public $rules = [
        'phone' => ['required', new ValidPhone], // Not valid php
    ];
}
```

The intent here is to allow the use of the `$rules` property if the user wants, but to also allow the `rules` method as a way to make the above possible.